rac2,replica_rac2: order replicaSendStream.mu before Replica.mu
The other way around is too hard to work with when we add pull mode. We
will call MakeMsgAppAndAssumeSent from inside a replicaSendStream and
will be adjusting replicaSendStream state before and after that call.
During that call Replica.mu needs to be held, since Raft state for this
follower (like Next) is being modified. It is not convenient to release
and reacquire replicaSendStream.mu just for this purpose.

Informs #130433

Epic: CRDB-37515

Release note: None