roachtest: reduce wal-failover flake around stall durations
Previously, we ran the risk of a prolonged (longer than 30/60s)
disk stall if the stall command took longer to run and return a value
back to us. This change moves the timer around to reduce the likelihood
we'd end up unintentionally stalling the node for more than 60s and
trip up the disk stall detector.

Fixes #131182.

Epic: none

Release note: None