server: new admin endpoint to resolve tenant name to id
This PR introduces a new admin endpoint `ResolveTenantName` to
resolve tenant name to tenant ID if the tenant name is valid.
This endpoint is mainly useful in initializing SQL servers in
"external" process mode. In "shared" process mode, both the
tenant ID and tenant name are anyway available.

This endpoint is intended to return the tenant ID for the tenant
the SQL server is started with (if it is valid).

Epic: CRDB-39093
Informs: #105340, #105339
Release note: None